requirements.txt
django
gunicorn
django-heroku
djangorestframework
dnspython
djongo
django-cors-headers
bs4
requests